+++ packages/component-library/src/components/Button/Button.tsx
@@ -1,66 +1,48 @@
-import React from "react";
-import styled from "styled-components";
-import { variant } from "styled-system";
+/** @jsxImportSource theme-ui */
 
-const variants = {
-  default: {
-    backgroundColor: "gray3",
+import React, { ComponentProps } from "react";
+import { ThemeUIStyleObject } from "theme-ui";
+
+const styles: ThemeUIStyleObject = {
+  borderRadius: "small",
+  border: "none",
+  padding: "small",
+  px: "medium",
+  fontSize: "large",
+  cursor: "pointer",
+  userSelect: "none",
+  display: "flex",
+  minHeight: "large",
+  alignItems: "center",
+  justifyContent: "center",
+  width: "max-content",
+
+  "&:active": {
+    opacity: 0.8,
   },
-  error: {
-    backgroundColor: "red7",
-  },
-  success: {
-    backgroundColor: "green7",
-  },
-  warning: {
-    backgroundColor: "orange6",
+
+  "&:disabled": {
+    cursor: "not-allowed",
+    opacity: "0.4",
   },
 };
 
-const StyledButton = styled.button<{ variant?: keyof typeof variants }>(
-  variant({ variants }),
-  ({ theme }) => ({
-    borderRadius: theme.radii.small,
-    borderStyle: "solid",
-    borderWidth: 2,
-    borderColor: theme.colors.gray9,
-    padding: `0 ${theme.space.medium}`,
-    fontSize: theme.fontSizes.large,
-    color: theme.colors.hiContrast,
-    cursor: "pointer",
-    userSelect: "none",
-    display: "flex",
-    minHeight: theme.space.xxl,
-    alignItems: "center",
-    justifyContent: "center",
-    width: "max-content",
-    boxShadow: `2px 2px 0 0px ${theme.colors.gray9}`,
-    transition: "transform, box-shadow, 0.2s ease-in-out",
-
-    "&:hover": {
-      transform: "translate(-1px, -1px)",
-      boxShadow: `4px 4px 8px 1px ${theme.colors.gray9}`,
-    },
-
-    "&:active": {
-      transform: "translate(2px, 2px)",
-      boxShadow: "none",
-    },
-
-    "&:disabled": {
-      cursor: "not-allowed",
-      opacity: "0.4",
-    },
-  })
-);
-
-export const Button = React.forwardRef<HTMLButtonElement>(
-  (props, forwardedRef) => {
-    return <StyledButton {...props} ref={forwardedRef} />;
-  }
-) as typeof StyledButton;
-
-Button.displayName = "Button";
-Button.defaultProps = {
-  variant: "default",
+const StyledButton = ({
+  variant,
+  children,
+}: {
+  variant?: string;
+  children: React.ReactNode;
+  ref?: React.ForwardedRef<HTMLButtonElement> | undefined;
+}) => {
+  return <button sx={{ ...styles, variant }}>{children}</button>;
 };
+
+export const Button = React.forwardRef<
+  HTMLButtonElement,
+  ComponentProps<typeof StyledButton>
+>((props, forwardedRef) => {
+  return <StyledButton {...props} ref={forwardedRef} />;
+});
+
+Button.displayName = "MiraButton";