+++ packages/design-tokens/src/index.ts
@@ -1,0 +1,8 @@
+export { default as colors } from "../tokens/colors";
+export { default as fontFamilies } from "../tokens/fontFamilies";
+export { default as fontSizes } from "../tokens/fontSizes";
+export { default as fontWeights } from "../tokens/fontWeights";
+export { default as mediaQueries } from "../tokens/mediaQueries";
+export { default as radii } from "../tokens/radii";
+export { default as shadows } from "../tokens/shadows";
+export { default as spacing } from "../tokens/spacing";